Editor: load data synchronously when data is stale
When the data is missing or stale the editor would show an empty list of
epics and sprints. Instead the editor should call the Jira API
synchronously in this instance too.

For this a LoadDataBlocking has been added that should load all
neccessary data concurrently but only return once all data has been
fetched.